Stop loading once a duplicate is found
If we go to create a database entry for a scrobble only to find it
already in the DB, then we’re up to date and can stop fetching more.

With an option to continue anyway, in case the original load of
all historic scrobbles failed to complete for some reason.